Added a profile feature for the homepage.
Profile for Lanyon is a Medium-esque area to display an avatar, site
title and a bio.

The bio is an additional option in the _config.yml file.

The addition of this required another layout called homepage.html (the
expectation is that this will be used on the homepage although this can
be changed).

This is no way impacts the current masthead. Instead it is a
replacement for the masthead on the homepage.